* 'hooks' of github.com:wireapp/wire-webapp: (74 commits)
  Improve center messages check
  Set picture should always return a Promise (#1336)
  Report sequential database update failures (#1335)
  Set client properties by default (#1334)
  Update documentation
  Add safeguard to disabling Localytics
  render_media_embeds: Node list is not an array
  Simplify preview length check
  Set proper scope for send_ping
  Add '_add_mediastream' failure case
  Change error on '_update_last_read_timestamp'
  Add '_add_mediastream' failure case
  Refactor rejection in _show_upload_warning
  Add safeguard to _focus_message
  Fix user picture update on Edge (#1333)
  Add safeguard to CollectionViewModel.set_conversation
  Add safeguard to CollectionViewModel.set_conversation
  Add check fur current device name
  Fix device removal name update
  Add safeguard to upload failed update (#1332)
  ...

+++ app/script/calling/v3/CallCenter.js
@@ -341,7 +341,7 @@ z.calling.v3.CallCenter = class CallCenter {
         .then((e_call_et) => this._confirm_e_call_message(e_call_et, e_call_message_et))
         .then((e_call_et) => e_call_et.delete_e_participant(user_id, client_id, termination_reason))
         .then(function(e_call_et) {
-          if (!e_call_et.is_group()) {
+          if (!e_call_et.is_group) {
             e_call_et.deactivate_call(e_call_message_et, termination_reason);
           }
         })
@@ -405,21 +405,13 @@ z.calling.v3.CallCenter = class CallCenter {
         e_call_et.set_remote_version(e_call_message_et);
 
         if (response) {
-          switch (e_call_et.state()) {
-            case z.calling.enum.CALL_STATE.INCOMING: {
-              this.logger.info(`Incoming e-call in conversation '${e_call_et.conversation_et.display_name()}' accepted on other device`);
-              return this.delete_call(conversation_id);
-            }
-
-            case z.calling.enum.CALL_STATE.OUTGOING: {
-              return e_call_et.update_e_participant(user_id, e_call_message_et)
-                .then(() => e_call_et.state(z.calling.enum.CALL_STATE.CONNECTING));
-            }
-
-            default: {
-              break;
-            }
+          if (user_id === this.user_repository.self().id) {
+            this.logger.info(`Incoming e-call in conversation '${e_call_et.conversation_et.display_name()}' accepted on other device`);
+            return this.delete_call(conversation_id);
           }
+
+          return e_call_et.update_e_participant(user_id, e_call_message_et)
+            .then(() => e_call_et.state(z.calling.enum.CALL_STATE.CONNECTING));
         }
 
         this.user_repository.get_user_by_id(user_id)
@@ -479,7 +471,7 @@ z.calling.v3.CallCenter = class CallCenter {
    * @returns {undefined} Resolves if the message is valid
    */
   _validate_message_destination(e_call_et, e_call_message_et) {
-    if (e_call_et.is_group()) {
+    if (e_call_et.is_group) {
       const {dest_client_id, dest_user_id, type} = e_call_message_et;
 
       if (dest_user_id !== this.user_repository.self().id || dest_client_id !== this.client_repository.current_client().id) {
